==> analysis/01_compare_freq_prev.Rmd <==
---
title: "Compare prev and allele freq SEA"
output: html_document
date: "2024-04-17"
---

```{r}
allele_freq_year <- read.table("data/data-derived/pfk7_data_freq_year.txt")
allele_freq <- read.table("data/data-derived/pfk7_data_freq.txt")
prev <- read.table("data/data-derived/pfk7_data.txt")
prev_WT <- read.table("data/data-derived/pfk7_data_WT.txt")
```

#Compare dataframes
```{r}
allele_freq_compare <- allele_freq %>%
  rename(n_freq = n)%>%
  select(country, District, year, Locus, n_freq, freq)
prev_compare <- prev %>%
  rename(n_prev = n) %>%
  select(country, District, year, Locus, n_prev, prev)

compare_df <- allele_freq_compare %>%
  group_by(country, District, year, Locus) %>%
  dplyr::left_join(prev_compare, join_by(country, District, year, Locus))
```

# Bar plot of prevalenc with respect to everything
```{r}
prev_barplot <- prev %>%
  group_by(year, Locus) %>%
  mutate(all_genotyped = sum(n, na.rm=TRUE), #all genotyped in a year per Locus
         mut_sample = sum(x, na.rm=TRUE))  %>% #mutations in a year per Locus
  ungroup %>%
  group_by(year) %>%
  mutate(all_genotyped = max(all_genotyped),
         n = length(unique(District))) %>%
  select(Locus, year, adj_year, n, all_genotyped, mut_sample) %>%
  distinct(Locus, all_genotyped, mut_sample, .keep_all = TRUE) %>%
  mutate(prev_all = mut_sample/all_genotyped) %>%
  ungroup() %>%
  ggplot(aes(x = year, y = prev_all, fill=Locus, color = Locus)) +
  geom_bar(stat = "identity", position = position_dodge2(preserve = "single")) +
  theme_bw() +
  scale_color_manual("Mutations",
    values = c("#56B4E9",
               "#009E73",
               "#BBBBBB",
               "#E69F00"),
    labels = c("580Y",
               "other K13",
               "K13",
               "539T")) +
  scale_fill_manual("Mutations",
    values = c("#56B4E9",
               "#009E73",
               "#BBBBBB",
               "#E69F00"),
    labels = c("580Y",
               "other K13",
               "K13",
               "539T")) +
  xlab("Year") +
  ylab("Prevalence") +
  ggtitle("Prevalence distribution of K13") +
  scale_y_continuous(labels = scales::percent) +
  scale_x_continuous("Year",
                     breaks = seq(2001, 2018, 2)) +
  theme(legend.title=element_blank(), 
        axis.text = element_text(size = 10), 
        axis.title=element_text(size=12), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 10), 
        legend.position = c(.2 , .99), 
        legend.justification = c("right", "top"), 
        legend.box.just = "right", 
        legend.margin = margin(1, 1, 1, 1))

ggsave("figures/REVISION/Prevalence_barplot.png",prev_barplot, dpi=400)
```

# Bar plot of prevalence with respect to WT
```{r}
prev_WT_barplot <- prev_WT %>%
  group_by(year, Locus) %>%
  mutate(mut_sample = sum(x, na.rm=TRUE),
         WT_sum = sum(WT, na.rm=TRUE))  %>% #mutations in a year per Locus
  ungroup %>%
  group_by(year) %>%
  mutate(WT_sum = max(WT_sum)) %>%
  select(Locus, year, adj_year, WT_sum, mut_sample) %>%
  distinct(Locus, mut_sample, .keep_all = TRUE) %>%
  mutate(prev_all = mut_sample/(mut_sample + WT_sum)) %>%
  ungroup() %>%
  ggplot(aes(x = year, y = prev_all, fill=Locus, color = Locus)) +
  geom_bar(stat = "identity", position = position_dodge2(preserve = "single")) +
  theme_bw() +
  scale_color_manual("Mutations",
    values = c("#56B4E9",
               "#009E73",
               "#BBBBBB",
               "#E69F00"),
    labels = c("580Y",
               "other K13",
               "K13",
               "539T")) +
  scale_fill_manual("Mutations",
    values = c("#56B4E9",
               "#009E73",
               "#BBBBBB",
               "#E69F00"),
    labels = c("580Y",
               "other K13",
               "K13",
               "539T")) +
  xlab("Year") +
  ylab("Prevalence") +
  ggtitle("Prevalence distribution of K13 over WT") +
  scale_y_continuous(labels = scales::percent) +
  scale_x_continuous("Year",
                     breaks = seq(2001, 2018, 2)) +
  theme(legend.title=element_blank(), 
        axis.text = element_text(size = 10), 
        axis.title=element_text(size=12), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 10), 
        legend.position = c(.2 , .99), 
        legend.justification = c("right", "top"), 
        legend.box.just = "right", 
        legend.margin = margin(1, 1, 1, 1))

ggsave("figures/REVISION/Prevalence_WT_barplot.png",prev_WT_barplot, dpi=400)
```

# Bar plot with respect to allele freq
```{r}
freq_barplot <- allele_freq_year %>%
  ggplot(aes(x = year, y = freq, fill=Locus, color = Locus)) +
  geom_bar(stat = "identity", position = position_dodge2(preserve = "single")) +
  theme_bw() +
  scale_color_manual("Mutations",
    values = c("#56B4E9",
               "#BBBBBB",
               "#009E73",
               "#E69F00"),
    labels = c("580Y",
               "K13",
               "other K13",
               "539T")) +
  scale_fill_manual("Mutations",
    values = c("#56B4E9",
               "#BBBBBB",
               "#009E73",
               "#E69F00"),
    labels = c("580Y",
               "K13",
               "other K13",
               "539T")) +
  xlab("Year") +
  ylab("Allele frequency") +
  ggtitle("Allele frequency distribution of K13 over WT") +
  #scale_y_continuous(labels = scales::percent) +
  scale_x_continuous("Year",  
                     breaks = seq(2001, 2018, 2)) +
  theme(legend.title=element_blank(), 
        axis.text = element_text(size = 10), 
        axis.title=element_text(size=12), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 10), 
        legend.position = c(.2 , .99), 
        legend.justification = c("right", "top"), 
        legend.box.just = "right", 
        legend.margin = margin(1, 1, 1, 1))

ggsave("figures/REVISION/Allele_freq_barplot.png",freq_barplot, dpi=400)
```
