==> analysis/03_Uganda_Modeling_prediction.Rmd <==
---
title: "Kelch Uganda Prediction for 2,3,4 years"
output: html_document
date: "2023-04-26"
---

```{r setup, include=FALSE}
library(lme4)
library(nlme)
library(dplyr)
library(ggplot2)
library(ggpubr)
library(scales)
library(tidyverse)
library(cowplot)
library(tabulizer)
library(RColorBrewer)
library(tidyverse) # ggplot2, dplyr, tidyr, readr, purrr, tibble
library(rnaturalearth)
library(rnaturalearthdata)
library(sf) #see ubuntu issues here: https://rtask.thinkr.fr/installation-of-r-4-0-on-ubuntu-20-04-lts-and-tips-for-spatial-packages/
library(ggspatial)
library(scatterpie)
library(scales)
library (ggforce)
library (ggmap)
library(rstan)
library(gt)
## install.packages("rstan", repos = c("https://mc-stan.org/r-packages/", getOption("repos")))

```

# Load data
```{r}
#load data
df <- read.table("data/data-derived/Ugandan_data.txt")
df_comb <- read.table("data/data-derived/Ugandan_comb_data.txt")
#set the limit of number of observations
numofobs <- 2

fileLocManuscript <- "figures/"

df <- df %>%
  group_by(District, Locus) %>%
  mutate(min_year = min(year[x > 0]),
         max_year = max(year[x > 0])) %>%
  ungroup 
```

# Forecasting selection up until 2023
```{r}
#------ Model prediction for C469Y ------
c469y_pred <- rstanarm::stan_glmer(
  lrsmed ~ adj_year + (1 + adj_year | District),
  data = df %>% 
    filter(Locus == "C469Y" & nobs > numofobs),
  weights = wt_med,
  adapt_delta = 0.9975
)

# forecasting for C469Y
new_data <- df %>%
  filter(nobs > numofobs, 
         is.finite(df$lrsmed), 
         Locus == "C469Y") %>% # this is for c469Y model
  group_by(District, Locus) %>% # we group by the District and Locus
  complete(year = seq(max(year), 2033, 1)) %>% # this creates all the year values up to 2023
  select(adj_year, year, District, Locus, min_year, max_year) %>%
  mutate(min_year = replace(min_year, is.na(min_year), 0),
         max_year = replace(max_year, is.na(max_year), 0)) 

new_data <- new_data %>%
    group_by(District, Locus) %>%
    mutate(min_year = max(min_year),
      adj_year = year - min_year) %>%
    arrange(Locus, District, year) %>%
    ungroup

new_data$predict <- rstanarm::posterior_predict(c469y_pred, type = "response", newdata = new_data) %>% colMeans
conf_val <- rstanarm::posterior_predict(c469y_pred, type = "response", newdata = new_data) %>% 
  apply(2, quantile, prob = (c(0.025, 0.975)))
new_data$conf.low <- conf_val[1,]
new_data$conf.high <- conf_val[2,] 

df_predict <- new_data

#------ Model prediction for A675V ------
a675v_pred <- rstanarm::stan_glmer(
  lrsmed ~ adj_year + (1 + adj_year | District),
  data = df %>% 
    filter(Locus == "A675V" & nobs > numofobs),
  weights = wt_med,
  adapt_delta = 0.9975
)

# forecasting for A467V
new_data <- df %>%
  filter(nobs > numofobs, 
         is.finite(df$lrsmed), 
         Locus == "A675V") %>% # this is for c469Y model
  group_by(District, Locus) %>% # we group by the District and Locus
  complete(year = seq(max(year), 2033, 1)) %>% # this creates all the year values up to 2023
  select(adj_year, year, District, Locus, min_year, max_year) %>%
  mutate(min_year = replace(min_year, is.na(min_year), 0),
         max_year = replace(max_year, is.na(max_year), 0)) 

new_data <- new_data %>%
    group_by(District, Locus) %>%
    mutate(min_year = max(min_year),
      adj_year = year - min_year) %>%
    arrange(Locus, District, year) %>%
    ungroup

new_data$predict <- rstanarm::posterior_predict(c469y_pred, type = "response", newdata = new_data) %>% colMeans
conf_val <- rstanarm::posterior_predict(c469y_pred, type = "response", newdata = new_data) %>% 
  apply(2, quantile, prob = (c(0.025, 0.975)))
new_data$conf.low <- conf_val[1,]
new_data$conf.high <- conf_val[2,] 

df_predict <- rbind(df_predict, new_data)

#------ Model prediction for C469F ------
c469f_pred <- rstanarm::stan_glmer(
  lrsmed ~ adj_year + (1 + adj_year | District),
  data = df %>% 
    filter(Locus == "C469F" , nobs > numofobs),
  weights = wt_med,
  adapt_delta = 0.9975
)


# forecasting for C469F
new_data <- df %>%
  filter(nobs > numofobs, is.finite(df$lrsmed), Locus == "C469F") %>% # this is for c469Y model
  group_by(District, Locus) %>% # we group by the District and Locus
  complete(year = seq(max(year), 2033, 1)) %>% # this creates all the year values up to 2023
  select(adj_year, year, District, Locus, min_year, max_year) %>%
  mutate(min_year = replace(min_year, is.na(min_year), 0),
         max_year = replace(max_year, is.na(max_year), 0)) 

new_data <- new_data %>%
    group_by(District, Locus) %>%
    mutate(min_year = max(min_year),
      adj_year = year - min_year) %>%
    arrange(Locus, District, year) %>%
    ungroup

new_data$predict <- rstanarm::posterior_predict(c469y_pred, type = "response", newdata = new_data) %>% colMeans
conf_val <- rstanarm::posterior_predict(c469y_pred, type = "response", newdata = new_data) %>% 
  apply(2, quantile, prob = (c(0.025, 0.975)))
new_data$conf.low <- conf_val[1,]
new_data$conf.high <- conf_val[2,] 

df_predict <- rbind(df_predict, new_data)

df_predict_allyear <- df_predict %>% 
  group_by(District, Locus) %>% 
  mutate(min_year = replace_na(min_year, 0),
         min_year = max(min_year),
         max_year = replace_na(max_year, 0),
         max_year = max(max_year))
```

## Convert to prevalence
```{r}
df_predict_allyear <- df_predict_allyear %>% mutate(predict_prev = exp(predict)*100/(1+exp(predict)),
                                                    conf.low_prev = exp(conf.low)*100/(1+exp(conf.low)),
                                                    conf.high_prev = exp(conf.high)*100/(1+exp(conf.high)))
```

## Figure 6: Ploting forecasting of predictions for the first 2, 3, and 4 years
```{r}
################ PLOT FOR A675V ################
df_true <- df_predict_allyear %>% 
  filter(Locus == "A675V", 
         year <= max_year) %>%
  mutate(label = "true")
df_pred <- df_predict_allyear %>% 
  filter(Locus == "A675V",
         year >= max_year) %>%
  mutate(label = "predicted")
ex_pt <- df %>% 
  filter(nobs > numofobs, Locus == "A675V", District %in% unique(df_pred$District))
  
pred_a675v <- ggplot() +
  geom_line(data = df_true, 
            aes(year, predict_prev, color = label)) +
  geom_line(data = df_pred, 
            aes(year, predict_prev, color = label)) +
  geom_ribbon(data = df_pred, 
              aes(ymin = conf.low_prev, ymax = conf.high_prev, x = year, color = label, fill = label), 
              linetype = "dashed", 
              alpha = 0.2,  
              show.legend = FALSE) +
  geom_point(data = ex_pt, 
             aes(x = year, y = med * 100, color = "true")) +
  theme_bw() +
  facet_wrap(~District) +
  ylab("A675V Prevalence") +
  xlab("Year") +
  scale_color_manual(
    values = c("#CC79A7", "grey"),
    labels = c("Forecasting",  "True")
  ) +
  scale_fill_manual(
    values = c("#CC79A7", "grey"),
    labels = c("Forecasting",  "True")
  ) +
  ggpubr::theme_pubclean(base_size = 12) +
  theme(
    axis.line = element_line(),
    axis.text = element_text(size = 6),
    axis.title = element_text(size = 10),
    strip.text.x = element_text(size = 10),
    legend.title = element_text(size = 10),
    legend.text = element_text(size = 8),
    legend.position = "none",
    panel.spacing = unit(0.2, "lines")) 

################ PLOT FOR C469Y ################
df_true <- df_predict_allyear %>% 
  filter(Locus == "C469Y", 
         year <= max_year) %>%
  mutate(label = "true")
df_pred <- df_predict_allyear %>% 
  filter(Locus == "C469Y",
         year >= max_year) %>%
  mutate(label = "predicted")
ex_pt <-df %>% 
  filter(nobs > numofobs, Locus == "C469Y", District %in% unique(df_pred$District)) 

pred_c469y <- ggplot() +
  geom_line(data = df_true, 
            aes(year, predict_prev, color = label)) +
  geom_line(data = df_pred, 
            aes(year, predict_prev, color = label)) +
  geom_ribbon(data = df_pred, 
              aes(ymin = conf.low_prev, ymax = conf.high_prev, x = year, color = label, fill = label), 
              linetype = "dashed", 
              alpha = 0.2,  
              show.legend = FALSE) +
  geom_point(data = ex_pt, 
             aes(x = year, y = med * 100, color = "true")) +
  theme_bw() +
  facet_wrap(~District) +
  ylab("C469Y Prevalence") +
  xlab("Year") +
  scale_color_manual(
    values = c("#CC79A7", "grey"),
    labels = c("Forecasting",  "True")
  ) +
  scale_fill_manual(
    values = c("#CC79A7", "grey"),
    labels = c("Forecasting",  "True")
  ) +
  ggpubr::theme_pubclean(base_size = 12) +
  theme(
    axis.line = element_line(),
    axis.text = element_text(size = 6),
    axis.title = element_text(size = 10),
    strip.text.x = element_text(size = 10),
    legend.title = element_text(size = 10),
    legend.text = element_text(size = 8),
    legend.position = "none",
    panel.spacing = unit(0.2, "lines")) 

################ PLOT FOR C469F ################
df_true <- df_predict_allyear %>% 
  filter(Locus == "C469F", 
         year <= max_year) %>%
  mutate(label = "true")
df_pred <- df_predict_allyear %>% 
  filter(Locus == "C469F",
         year >= max_year) %>%
  mutate(label = "predicted")
ex_pt <-df %>% 
  filter(nobs > numofobs, Locus == "C469F", District %in% unique(df_pred$District)) 

pred_c469f <- ggplot() +
  geom_line(data = df_true, 
            aes(year, predict_prev, color = label)) +
  geom_line(data = df_pred, 
            aes(year, predict_prev, color = label)) +
  geom_ribbon(data = df_pred, aes(ymin = conf.low_prev, ymax = conf.high_prev, x = year, color = label, fill = label), 
              linetype = "dashed", 
              alpha = 0.2,  
              show.legend = FALSE) +
  geom_point(data = ex_pt, 
             aes(x = year, y = med * 100, color = "true")) +
  theme_bw() +
  facet_wrap(~District) +
  ylab("C469F Prevalence") +
  xlab("Year") +
  scale_color_manual(
    values = c("#CC79A7", "grey"),
    labels = c("Forecasting",  "True")
  ) +
  scale_fill_manual(
    values = c("#CC79A7", "grey"),
    labels = c("Forecasting",  "True")
  ) +
  ggpubr::theme_pubclean(base_size = 12) +
  theme(
    axis.line = element_line(),
    axis.text = element_text(size = 6),
    axis.title = element_text(size = 10),
    strip.text.x = element_text(size = 10),
    legend.title = element_text(size = 10),
    legend.text = element_text(size = 8),
    legend.position = "none",
    panel.spacing = unit(0.2, "lines")) 
```

## Saving plots
```{r}
fig6_width = 6
fig6_height = 4
ggsave(
  "figures/Figure6-Forecasting/Uganda/Uganda_forecasting_A675V.png",
  pred_a675v,
  width = fig6_width,
  height = fig6_height,
  dpi = 700
)
ggsave(
  "figures/Figure6-Forecasting/Uganda/Uganda_forecasting_C469Y.png",
  pred_c469y,
  width = fig6_width,
  height = fig6_height,
  dpi = 700
)
ggsave(
  "figures/Figure6-Forecasting/Uganda/Uganda_forecasting_C469F.png",
  pred_c469f,
  width = fig6_width,
  height = 2.7,
  dpi = 700
)

ggsave(
  "figures/Figure6-Forecasting/Uganda/Uganda_forecasting_A675V.pdf",
  pred_a675v,
  width = fig6_width,
  height = fig6_height,
  dpi = 700
)
ggsave(
  "figures/Figure6-Forecasting/Uganda/Uganda_forecasting_C469Y.pdf",
  pred_c469y,
  width = fig6_width,
  height = fig6_height,
  dpi = 700
)
ggsave(
  "figures/Figure6-Forecasting/Uganda/Uganda_forecasting_C469F.pdf",
  pred_c469f,
  width = fig6_width,
  height = 2.7,
  dpi = 700
)
```
