==> analysis/05_Supplement_Uganda_DataOverview.Rmd <==
---
title: "Supplement -Klech Selection Coefficient Modeling"
output: html_document
date: "2023-05-06"
---

```{r setup, include=FALSE}
library(lme4)
library(nlme)
library(dplyr)
library(ggplot2)
library(ggpubr)
library(scales)
library(tidyverse)
library(cowplot)
library(tabulizer)
library(RColorBrewer)
library(tidyverse) # ggplot2, dplyr, tidyr, readr, purrr, tibble
library(rnaturalearth)
library(rnaturalearthdata)
library(sf) #see ubuntu issues here: https://rtask.thinkr.fr/installation-of-r-4-0-on-ubuntu-20-04-lts-and-tips-for-spatial-packages/
library(ggspatial)
library(scatterpie)
library(scales)
library (ggforce)
library (ggmap)
library(rstan)
library(gt)
library(ggbreak) 
## install.packages("rstan", repos = c("https://mc-stan.org/r-packages/", getOption("repos")))
```

# Load data
```{r}
#load data
df_uganda <- read.table("data/data-derived/Ugandan_data.txt")
df_pfk7 <- read.table("data/data-derived/pfk7_data.txt")
#set the limit of number of observations
numofobs <- 2

fileLocManuscript <- "figures/"
```

# Overview of data plot
## Uganda
```{r}
plot_overview_uganda <- df_uganda %>%
  mutate(N = n + x) %>%
  ggplot(aes(year, med * 100, group = Locus, color = Locus)) +
  geom_point() +
  labs(shape = "Mutations", size = "Sample Size") +
  theme_bw() +
  facet_wrap( ~ District) +
  ylab("Prevalence") +
  xlab("Year") +
  ylim(-1, 40) +
  scale_color_manual("Mutations",
    values = c("#D55E00", "#0072B2", "#009E73", "#F0E442"),
    labels = c("675V", "469F", "469Y", "561H")
  ) +
  ggpubr::theme_pubclean(base_size = 12) +
  theme(
    axis.line = element_line(),
    axis.text = element_text(size = 8),
    axis.title = element_text(size = 12),
    legend.text = element_text(size = 7),
    strip.text.x = element_text(size = 12),
    legend.title = element_text(size = 7),
    legend.position = c(0.94, 0.07),
    legend.box = "horizontal"
  )
```

### Saving plot
```{r}
ggsave(
  paste0(fileLocManuscript, "/supplement/Uganda_dataoverview.pdf"),
  plot_overview_uganda,
  width = 8,
  height = 5,
  dpi = 700
)
ggsave(
  paste0(fileLocManuscript, "/supplement/Uganda_dataoverview.png"),
  plot_overview_uganda,
  width = 8,
  height = 5,
  dpi = 700
)
```

## SE Asia (PfK7 data)
```{r}
d1 <- df_pfk7 %>% select(country, District, year, c580y, nobs_c580y, n, prev_580y) %>% mutate(Locus = "C580Y") %>% rename(x = c580y, nobs = nobs_c580y, prev = prev_580y)
d2 <- df_pfk7 %>% select(country, District, year, r539t, nobs_r539t, n, prev_539t) %>% mutate(Locus = "R539T") %>% rename(x = r539t, nobs = nobs_r539t, prev = prev_539t)
d3 <- df_pfk7 %>% select(country, District, year, all_kelch, nobs_kelch, n, prev_all_kelch) %>% mutate(Locus = "all_kelch") %>% rename(x=all_kelch, nobs = nobs_kelch, prev = prev_all_kelch)
prev_df <- rbind(d1, d2, d3)

plot_overview_seAsia <- prev_df %>%
  mutate(N = n + x) %>%
  ggplot(aes(year, prev * 100, group = Locus, color = Locus)) +
  geom_point() +
  labs(shape = "Mutations", size = "Sample Size") +
  theme_bw() +
  facet_wrap( ~ country + District) +
  ylab("Prevalence") +
  xlab("Year") +
  ylim(-1, 40) +
  scale_color_manual(values = c("#E69F00", "#56B4E9", "#009E73"), 
                     labels = c("PfK13", "580Y", "539T"),
                     name = "Mutations")  +
  ggpubr::theme_pubclean(base_size = 12) +
  theme(
    axis.line = element_line(),
    axis.text = element_text(size = 8),
    axis.title = element_text(size = 12),
    legend.text = element_text(size = 10),
    strip.text.x = element_text(size = 8),
    legend.title = element_text(size = 10),
    legend.position = c(0.9, 0.08),
    legend.box = "horizontal"
  )
```

### Saving plot
```{r}
ggsave(
  paste0(fileLocManuscript, "/supplement/SEAsia_dataoverview.pdf"),
  plot_overview_seAsia,
  width = 8,
  height = 10,
  dpi = 700
)
ggsave(
  paste0(fileLocManuscript, "/supplement/SEAsia_dataoverview.png"),
  plot_overview_seAsia,
  width = 8,
  height = 10,
  dpi = 700
)
```

# SE Asia
## Prevalence plot overview
```{r}
d1 <-
  df_pfk7 %>% select(country, District, year, c580y, nobs_c580y, n) %>% mutate(Locus = "C580Y") %>% rename(x = c580y, nobs = nobs_c580y)
d2 <-
  df_pfk7 %>% select(country, District, year, r539t, nobs_r539t, n) %>% mutate(Locus = "R539T") %>% rename(x = r539t, nobs = nobs_r539t)
d3 <-
  df_pfk7 %>% select(country, District, year, all_kelch, nobs_kelch, n) %>% mutate(Locus = "all_kelch") %>% rename(x =
                                                                                                                       all_kelch, nobs = nobs_kelch)
d3$x <- d3$x - d2$x - d1$x

prev_df <- rbind(d1, d2, d3)
prev_df <- prev_df %>% group_by(year, Locus) %>%
  mutate(min_n = sum(n), mut_sample = sum(x)) %>%
  distinct(min_n, mut_sample) %>%
  mutate(prop_sample = mut_sample / min_n)

prev <- prev_df %>%
  ggplot(aes(x = year, y = prop_sample, fill = Locus)) +
  geom_col() +
  theme_bw() +
  scale_fill_manual("Mutations",
    values = c("#E69F00", "#56B4E9", "#009E73"),
    labels = c("other PfK13", "580Y", "539T")
  ) +
  xlab("Year") +
  ylab("Allele frequency") +
  ylim(0, 1) +
  scale_x_continuous(breaks = seq(1993, 2018, 3)) +
  theme(
    legend.title = element_blank(),
    axis.text = element_text(size = 10),
    axis.title = element_text(size = 12),
    legend.text = element_text(size = 10),
    strip.text.x = element_text(size = 10),
    legend.position = c(0.18, 0.9)
  )
```

### Saving plot
```{r}
ggsave(
  paste0(fileLocManuscript, "/supplement/SEAsia_prevalence.pdf"),
  prev,
  width = 4.5,
  height = 6.25,
  dpi = 700
)
ggsave(
  paste0(fileLocManuscript, "/supplement/SEAsia_prevalence.png"),
  prev,
  width = 4.5,
  height = 6.25,
  dpi = 700
)
```

## Download and preprocessing map of SE Asia
```{r}
#load base map layers
admin8 <-
  ne_download(
    scale = "large",
    type = "admin_1_states_provinces_lines",
    category = "cultural",
    returnclass = "sf"
  )
rivers8 <- ne_download(
  scale = 10,
  type = 'rivers_lake_centerlines',
  category = 'physical',
  returnclass = "sf"
)
lakes8 <- ne_download(
  scale = "large",
  type = 'lakes',
  category = 'physical',
  returnclass = "sf"
)
sov110 <- ne_download(
  scale = "medium",
  type = "sovereignty",
  category = "cultural",
  returnclass = "sf"
)
admin110 <- ne_download(
  scale = "large",
  type = "populated_places",
  category = "cultural",
  returnclass = "sf"
)

pt.lim = data.frame(xlim = c(28.5, 35.5), ylim = c(4.25, -3))
pt.bbox <- st_bbox(c(
  xmin = pt.lim$xlim[1],
  xmax = pt.lim$xlim[2],
  ymin = pt.lim$ylim[1],
  ymax = pt.lim$ylim[2]
))

SE_asia <-
  ne_states(country = unique(df_pfk7$country),
            returnclass = "sf") %>%
  #st_crop(pt.bbox) %>%
  select(name, admin, latitude, longitude) %>%
  arrange(name) %>% arrange(admin) %>%
  mutate(
    name = str_replace(name, "Batdâmbâng", "Battambang"),
    name = str_replace(name, "Kaôh Kong", "Koh Kong"),
    name = str_replace(name, "Krong Pailin", "Pailin"),
    name = str_replace(name, "Preah Vihéar", "Preah Vihear"),
    name = str_replace(name, "Pouthisat", "Pursat"),
    name = str_replace(name, "Si Sa Ket", "Sisakhet"),
    name = str_replace(name, "Rôtânôkiri", "Ratanakiri"),
    name = str_replace(name, "Attapu", "Attapeu"),
    name = str_replace(name, "Saravan", "Salavan"),
    name = str_replace(name, "Savannakhét", "Savannakhet"),
    name = str_replace(name, "Xékong", "Sekong"),
    name = str_replace(name, "Khánh Hòa", "Khanh Hoa"),
    name = str_replace(name, "Quàng Nam", "Quang Nam")
  )

SE_asia$name[SE_asia$latitude == 9.33941] <- "Bac Lieu"
SE_asia$name[SE_asia$latitude == 11.6805] <- "Binh Phuoc"
SE_asia$name[SE_asia$latitude == 11.1284] <- "Binh Thuan"
SE_asia$name[SE_asia$latitude == 12.7691] <- "Dak Lak"
SE_asia$name[SE_asia$latitude == 12.1425] <- "Dak Nong"
SE_asia$name[SE_asia$latitude == 11.6242] <- "Ninh Thuan"
SE_asia$name[SE_asia$latitude == 16.7204] <- "Quang Tri"
SE_asia$name[SE_asia$latitude == 13.6662] <- "Stueng Traeng"

SE_asia <-
  SE_asia  %>% filter(name %in% unique(df_pfk7$District)) %>% na.omit() %>% arrange(name)

SE_asia$color <- "not_enough_samples"
r539t_district <-
  df_pfk7 %>% filter(is.finite(lrsmed_539t), nobs_r539t > numofobs) %>% distinct(District)
c580y_district <-
  df_pfk7 %>% filter(is.finite(lrsmed_580y), nobs_c580y > numofobs) %>% distinct(District)
allkelch_district <-
  df_pfk7 %>% filter(is.finite(lrsmed_all_kelch), nobs_kelch > numofobs) %>% distinct(District)

r539t_district <- t(as.data.frame(r539t_district))
c580y_district <- t(as.data.frame(c580y_district))
allkelch_district <- t(as.data.frame(allkelch_district))

districts <- cbind(c580y_district, r539t_district)

allkelch_district <-
  allkelch_district[!allkelch_district %in% districts]
r539t_district <- array(r539t_district)
c580y_district <- array(c580y_district)

SE_asia$color[SE_asia$name %in% r539t_district] <- "r539t" #combined
SE_asia$color[SE_asia$name %in% c580y_district] <-
  "c580y, " #combined
SE_asia$color[SE_asia$name %in% allkelch_district] <-
  "other PfK13" #only other kelch mutants

pie_data <-
  df_pfk7 %>% mutate(
    WildType = n - c580y - r539t - (all_kelch - c580y - r539t),
    all_kelch = all_kelch - c580y - r539t
  ) %>% select(District, n, c580y, r539t, all_kelch, WildType)

pie_data <-
  pie_data %>% group_by(District) %>% mutate(
    c580y = sum(c580y),
    r539t = sum(r539t),
    all_kelch = sum(all_kelch),
    WildType = sum(WildType)
  ) %>% arrange(District) %>% ungroup %>% distinct(District, .keep_all = TRUE)
pie_data$lat <- NA
pie_data$long <- NA
for (i in unique(pie_data$District)) {
  pie_data$lat[pie_data$District == i] <-
    rep(SE_asia$latitude[SE_asia$name == i], sum((pie_data$District == i)))
  pie_data$long[pie_data$District == i] <-
    rep(SE_asia$longitude[SE_asia$name == i], sum((pie_data$District == i)))
}

pie_data$radius <-
  (pie_data$c580y + pie_data$r539t + pie_data$all_kelch) / pie_data$WildType
minr <- min(pie_data$radius)  # min in the data original scale
maxr <- max(pie_data$radius)  # max in the data original scale

min_scale <- 0.3  ## min in the new scale
max_scale <- 0.7 ## max in the new scale

pie_data$radius_adj <-
  min_scale + (max_scale - min_scale) * pie_data$radius / (maxr - minr)
```

### Map with pie chart
```{r}
# make plot with wild_type
map_SEAsia <- ggplot() +
  geom_sf(
    data = sov110,
    color = 'black',
    size = 0.3,
    fill = ifelse(sov110$ADMIN %in% c("Laos", "Thailand", "Cambodia", "Vietnam", "Myanmar"), 'grey98', 'grey90')
  ) +
  geom_sf(data = SE_asia, fill = "moccasin") +
  geom_sf(
    data = lakes8,
    color = "grey40",
    fill = "aliceblue",
    size = 0.5
  ) +
  geom_sf(data = admin8, color = "grey80", size = 0.6) +
  geom_sf(data = admin110, color = "grey80", size = 0.4) +
  annotation_scale(
    location = "bl",
    width_hint = 0.3,
    height = unit(0.1, "cm")
  ) +
  annotation_north_arrow(
    location = "bl",
    which_north = "true",
    pad_x = unit(0.25, "in"),
    pad_y = unit(0.15, "in"),
    style = north_arrow_fancy_orienteering,
    height = unit(1, "cm"),
    width = unit(1, "cm")
  ) +
  coord_sf(xlim = c(109, 92.5),
           ylim = c(6.5, 28),
           expand = TRUE) +
  annotate(
    "text",
    x = 102.5,
    y = 20,
    label = "Laos",
    color = "black",
    size = 3 ,
    fontface = "bold"
  ) +
  annotate(
    "text",
    x = 101.5,
    y = 16,
    label = "Thailand",
    color = "black",
    size = 3 ,
    fontface = "bold"
  ) +
  annotate(
    "text",
    x = 105.9,
    y = 12.8,
    label = "Cambodia",
    color = "black",
    size = 3 ,
    fontface = "bold"
  ) +
  annotate(
    "text",
    x = 104.9,
    y = 21.5,
    label = "Vietnam",
    color = "black",
    size = 3 ,
    fontface = "bold"
  ) +
  annotate(
    "text",
    x = 95.2,
    y = 16.9,
    label = "Myanmar",
    color = "black",
    size = 3 ,
    fontface = "bold"
  ) +
  annotate(
    "text",
    x = 102,
    y = 24,
    label = "China",
    color = "grey50",
    size = 4 ,
    fontface = "italic"
  ) +
  theme_minimal() +
  xlab("") +
  ylab("") +
  geom_scatterpie(
    data = pie_data,
    aes(
      x = long,
      y = lat,
      group = District,
      r = radius_adj
    ),
    cols = c("c580y", "r539t", "all_kelch")
  ) +
  geom_scatterpie_legend(
    pie_data$radius_adj*2,
    x = 106,
    y = 24,
    labeller = function(x)
      round(100*(34*x/2 - 10.2))
  ) +
  scale_fill_manual(values = c("#56B4E9", "#009E73", "#E69F00"), labels = c("580Y", "539T", "other PfK13"), name = "")  +
  theme(
    legend.text = element_text(size = 10),
    axis.text = element_text(size = 10),
    legend.position = c(0.86, 0.95)
  )
```

#### Saving plot
```{r}
ggsave(
  paste0(
    fileLocManuscript,
    "/supplement/SEAsia_Map.pdf"
  ),
  map_SEAsia,
  width = 5.625,
  height = 6.25,
  dpi = 700
)
ggsave(
  paste0(
    fileLocManuscript,
    "/supplement/SEAsia_Map.png"
  ),
  map_SEAsia,
  width = 5.625,
  height = 6.25,
  dpi = 700
)
```


### Map of SE Asia
```{r}
# make plot with wild_type
map_SEAsia <- ggplot() +
  geom_sf(
    data = sov110,
    color = 'black',
    size = 0.3,
    fill = ifelse(
      sov110$ADMIN %in% c("Laos", "Thailand", "Cambodia", "Vietnam", "Myanmar"),
      'grey98',
      'grey90'
    )
  ) +
  geom_sf(data = SE_asia, aes(fill = color)) +
  geom_sf(
    data = lakes8,
    color = "grey40",
    fill = "aliceblue",
    size = 0.5
  ) +
  geom_sf(data = admin8, color = "grey80", size = 0.6) +
  geom_sf(data = admin110, color = "grey80", size = 0.4) +
  annotation_scale(
    location = "bl",
    width_hint = 0.3,
    height = unit(0.1, "cm")
  ) +
  annotation_north_arrow(
    location = "bl",
    which_north = "true",
    pad_x = unit(0.25, "in"),
    pad_y = unit(0.15, "in"),
    style = north_arrow_fancy_orienteering,
    height = unit(1, "cm"),
    width = unit(1, "cm")
  ) +
  coord_sf(xlim = c(109, 91),
           ylim = c(6.5, 28),
           expand = TRUE) +
  annotate(
    "text",
    x = 102.5,
    y = 20,
    label = "Laos",
    color = "black",
    size = 3 ,
    fontface = "bold"
  ) +
  annotate(
    "text",
    x = 101.5,
    y = 16,
    label = "Thailand",
    color = "black",
    size = 3 ,
    fontface = "bold"
  ) +
  annotate(
    "text",
    x = 105.9,
    y = 12.8,
    label = "Cambodia",
    color = "black",
    size = 3 ,
    fontface = "bold"
  ) +
  annotate(
    "text",
    x = 104.9,
    y = 21.5,
    label = "Vietnam",
    color = "black",
    size = 3 ,
    fontface = "bold"
  ) +
  annotate(
    "text",
    x = 95.2,
    y = 16.9,
    label = "Myanmar",
    color = "black",
    size = 3 ,
    fontface = "bold"
  ) +
  annotate(
    "text",
    x = 106,
    y = 24,
    label = "China",
    color = "grey50",
    size = 4 ,
    fontface = "italic"
  ) +
  theme_minimal() +
  xlab("") +
  ylab("") +
  #geom_scatterpie(data = pie_data, aes(x=long, y=lat, group=District, r = radius_adj), cols=c("c580y", "r539t", "all_kelch")) +
  #geom_scatterpie_legend(pie_data$radius*2, x=28.7, y=2.5, labeller = function(x) round(x * 100/2)) +
  scale_fill_manual(
    values = c("#56B4E9", "moccasin", "#E69F00", "#009E73"),
    labels = c("580Y", "not enough samples", "other PfK13", "539T"),
    name = ""
  )  +
  theme(
    legend.text = element_text(size = 10),
    axis.text = element_text(size = 10),
    legend.position = c(0.77, 0.90)
  )

```
